move `flow_in_progress` into `dataflow` and document it
