+++ src/librustc_mir/borrow_check/nll/constraint_generation.rs
@@ -22,8 +22,8 @@ use rustc::ty::fold::TypeFoldable;
 use rustc::util::common::ErrorReported;
 use rustc_data_structures::fx::FxHashSet;
 use syntax::codemap::DUMMY_SP;
-use borrow_check::FlowInProgress;
 use dataflow::MaybeInitializedLvals;
+use dataflow::flow_in_progress::FlowInProgress;
 use dataflow::move_paths::{HasMoveData, MoveData};
 
 use super::LivenessResults;
@@ -172,7 +172,6 @@ impl<'cx, 'cg, 'gcx, 'tcx> ConstraintGeneration<'cx, 'cg, 'gcx, 'tcx> {
                         "add_liveness_constraints: location={:?} initialized={:?}",
                         location,
                         &self.flow_inits
-                            .base_results
                             .operator()
                             .move_data()
                             .move_paths[mpi_init]
