+++ src/librustc_mir/dataflow/mod.rs
@@ -35,6 +35,7 @@ pub(crate) use self::drop_flag_effects::*;
 use self::move_paths::MoveData;
 
 mod drop_flag_effects;
+pub(crate) mod flow_in_progress;
 mod graphviz;
 mod impls;
 pub mod move_paths;
